chg:usr:NEZ-100 Cache Materialized service
There are 2 impacts of this:
1. Parser can be cached now, as it caches MaterializedService.
Earlier every new thread would return new MaterializedService and that
caused issues with optimization.
2. As Materialized Service is now cached, populating materialized views
can be optimized without the need to parse queries that defines views.
With cubes number of views can be in thousands were this optimization
can significantly reduce the time to pupulate materializations.