Cache Materialized service for a Parser Object
Materialized service returns new instance for every thread.
So if we cache Parser object and another thread (not same as the thread
that created Parser) uses it to parse a query, we would get new MS
instance which needs to have all the materialized views registered
again. With this change we cache the materialized service instance and
use it for all the threads accessing same Parser object.