chg:usr:Accept java.sql.Types instead of strings for data types.
Previously QuarkColumn accepted types as a string. The accepted
strings were maintained in FieldTypes. It also mapped it to Java Type.
Different plugins had different strings for the same type and these
had to mapped to the known strings in FieldTypes.
This commit cleans up the implementation. FieldTypes is standardized on
java.sql.Types. These types are mapped to Java Types as per the standard.
Plugins are expected to provide java.sql.Types instead of strings when
constructing a QuarkColumn