new:usr:#9 Generic JDBC Plugin for unsupported databases
Quark contains a small list of databases that support JDBC such as
MySQL & RedShift. It is not possible to try Quark with a database
outside this list unless a new class is added. This commit add support for
accessing databases with a JDBC client by providing all the necessary information
through the JSON Catalog. GenericDB is not yet supported in DB Catalog.