new:usr: #20 Quark Server and Thin JDBC Client
Quark is available as a JDBC Server and a thin JDBC client. Calcite's
Avatica framework is used to implement the server and client. The client
is very small in size and can be easily distributed. The server accepts
a JSON configuration and connects to a JSON/DB Catalog.

JDBC directories and packages have been renamed to fatjdbc.
The new client is the official JDBC driver going forward. It uses the
jdbc namespace.