chg:usr:#60 Add optimization of joined tables using materialized views (#61)
There are couple of big changes related to this:
1. Introduces rule to optimize a join table with materialized
view. Rule waits for the predicate to be pushed onto join table and then
check if the predicate on the table can be satisifed by existing
materialized view in the system.
2. Formatting the output for Hive. Above rule introduces join on
subqueries which are not supported by Hive. So after optimization, for
hive we try to pull predicates out of the join and transform the subquery
joins to table join. Infact join ordering rule is also excluded from
optimization as that was also resulting in such subquery join what can
fail in Hive.