new:dev:#70:Encryption of database fields is optional. Encryptor is not a singleton
Accept a new field in credential -> "encrypt" to decide whether fields in the
DB catalog like url and password are encrypted or not.

MySQLAES was a singleton to allow passing of keys to JDBI mappers. The
problem with this approach is that its not possible to have multiple instances
of DB catalog in the same JVM. Use DBI::define to pass an Encrypt object instead.
This object is available in StatementContext argument of the map function in a
mapper.

Fix #70