new:usr:#69:Accept DSSet ID to support multi-tenant implementations. (#72)
DS Set is a group of data sources and their cubes and views. Typically in a
single tenant system there is only set and DSSets are not exposed to users
and admins. For multi-tenant systems allow connections to be created for a
particular DSSet. This argument is valid for DB Catalogs only and is optional.